created new window_management file will add more functions to manage/dynamically manage windows within an instance
